+++ src/Constants.java
@@ -3,7 +3,6 @@ public final class Constants {
     public static final String UKRAINIAN_LETTERS = "абвгґдеєжзиіїйклмнопрстуфхцчшщьюяАБВГҐДЕЄЖЗИІЇЙКЛМНОПРСТУФХЦЧШЩЬЮЯ";
     public static final String ENGLISH_LETTERS = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
     public static final String PUNCTUATION = ".,\":-!?( ”)";
-    public static final String CHARS = "!\"#$%&\\'()*+,-./0123456789:;<=>?@\\\\^_`{|}~ ";
     public static final String WRONG_NUMBER = "Invalid input! Please enter a valid number... 1, 2 or 3";
 
     private Constants() {
