+++ src/FileOpenSaveAndConvert.java
@@ -7,18 +7,18 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-public class FileConvertToChars {
+public class FileOpenSaveAndConvert {
 
     private String fileNameOpen;
     private String fileNameSave;
 
-    public FileConvertToChars(String fileName, String fileNameSave) {
-        this.fileNameOpen = fileName;
-        this.fileNameSave = fileNameSave;
+    public FileOpenSaveAndConvert() {
+
     }
 
-    public FileConvertToChars() {
-
+    public FileOpenSaveAndConvert(String fileName, String fileNameSave) {
+        this.fileNameOpen = fileName;
+        this.fileNameSave = fileNameSave;
     }
 
     public char[] bytesToChars() {
