+++ src/Main.java
@@ -3,11 +3,5 @@ public class Main {
     public static void main(String[] args) {
 
         new ConsoleMenu().startApp();
-//        FileConvertToChars textFile = new FileConvertToChars("src\\files\\TextUKR.txt", "src\\filesEncrypt\\EncryptTextUKR.txt");
-//        src\\filesDecrypt\\DecryptTextUKR.txt
-//
-//        EncryptDecryptText encryptText = new EncryptDecryptText();
-//        encryptText.encrypt(textFile);
-
     }
 }
