+++ src/keys/Key.java
@@ -1,0 +1,4 @@
+package keys;
+
+public class Key {
+}
