+++ src/keys/KeyOne.java
@@ -1,0 +1,4 @@
+package keys;
+
+public class KeyOne extends Key{
+}
