+++ src/keys/KeyThree.java
@@ -1,0 +1,4 @@
+package keys;
+
+public class KeyThree extends Key{
+}
