+++ src/keys/KeyTwo.java
@@ -1,0 +1,4 @@
+package keys;
+
+public class KeyTwo extends Key{
+}
